requirements.txt
aiofile==3.8.8
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
caio==0.9.13
certifi==2023.7.22
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.104.1
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
importlib-resources==6.1.0
Jinja2==3.1.2
limits==3.6.0
MarkupSafe==2.1.3
nodeenv==1.8.0
packaging==23.2
passlib==1.7.4
prisma==0.11.0
prometheus-client==0.18.0
prometheus-fastapi-instrumentator==6.1.0
psycopg2-binary==2.9.9
pyasn1==0.5.0
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
rsa==4.9
six==1.16.0
slowapi==0.1.8
sniffio==1.3.0
starlette==0.27.0
tomlkit==0.12.1
typing_extensions==4.8.0
uvicorn==0.23.2
wrapt==1.15.0
